feat: exception handling, introduce LockedException

+++ src/main/java/com/aj/flightregister/service/AirportService.java
@@ -2,8 +2,10 @@ package com.aj.flightregister.service;
 
 import com.aj.flightregister.exception.ItemAlreadyExistsException;
 import com.aj.flightregister.exception.ItemNotFoundException;
+import com.aj.flightregister.exception.LockedException;
 import com.aj.flightregister.model.Airport;
 import com.aj.flightregister.repository.AirportRepository;
+import com.aj.flightregister.repository.FlightRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
@@ -14,6 +16,8 @@ import java.util.List;
 public class AirportService {
 
     private final AirportRepository airportRepository;
+
+    private final FlightRepository flightRepository;
 
     public List<Airport> getAirports() {
         return airportRepository.findAll();
@@ -31,19 +35,24 @@ public class AirportService {
         } else throw new ItemAlreadyExistsException("Conflict", airport.getName());
     }
 
-    public void deleteAirport(String id) throws ItemNotFoundException {
+    public void deleteAirport(String id) throws ItemNotFoundException, LockedException {
+        Airport airport = airportRepository.findByName(id).orElseThrow(
+                () -> new ItemNotFoundException("Not found", id));
+        if (!flightRepository.findByOrigin_NameOrDestination_Name(airport.getName(), airport.getName()).isEmpty()) {
+            throw new LockedException("Foreign key", id);
+        }
         airportRepository.delete(
                 airportRepository
                         .findByName(id)
                         .orElseThrow(() -> new ItemNotFoundException("Not found", id)));
     }
 
-    public Airport updateAirport(Airport newAirport) {
+    public Airport updateAirport(Airport newAirport) throws ItemNotFoundException {
         return airportRepository.findByName(newAirport.getName()).map(
                 oldAirport -> {
                     Airport updated = oldAirport.updateWith(newAirport);
                     return airportRepository.save(updated);
-                }).get();
+                }).orElseThrow(() -> new ItemNotFoundException("Not found", newAirport.getName()));
 
     }
 }
